Campaign Updated

+++ app/api/contacts/(contact-main)/remove/route.ts
@@ -12,18 +12,28 @@ export async function POST(req: NextRequest) {
         return NextResponse.json({ error: 'User not authenticated' }, { status: 401 });
     }
 
+
+
     try {
+        console.log(id);
+        const contactId = parseInt(id, 10);
+
+        if (isNaN(contactId)) {
+            return NextResponse.json({ error: 'Invalid Contact Id', }, { status: 500 });
+        }
 
         const contact = await prisma.contact.delete({
-            where:{
-                id:+id
+            where: {
+                id: contactId
             }
         });
+        console.log(contact);
 
         return NextResponse.json({ msg: 'Contact Removed Successfully' }, { status: 200 });
 
     } catch (e) {
-        return NextResponse.json({ error: 'Something Went Wrong' }, { status: 500 });
+        // console.log(e);
+        return NextResponse.json({ error: 'Something Went Wrong', e: e }, { status: 500 });
     }
 
 }
